feat(gates): extract transcript analyzer to library
Move classification, bucketing, JSONL extraction, and rendering out of
scripts/analyze_yolo_transcripts.py into spellbook/gates/transcript_analyzer.py
so the same logic backs both the CLI script and the upcoming
permissions-from-transcripts skill. The script becomes a thin CLI wrapper
that re-exports library names for backward compatibility.

Tighten classification breadth (deferred from WI-2 review):

- Add per-subcommand 3-word runners for gh pr / gh run / gh issue / gh repo,
  splitting read-only forms (view, list, diff, status, checks, watch) into
  SEARCH_INSPECT and mutating forms (create, edit, review, ready, close,
  merge, reopen, rerun, cancel, delete) into MUTATING.
- Add 4-word runners for acli jira workitem (view/list as read-only,
  transition/edit as mutating).
- Leave gh api UNCLASSIFIED with a code comment: it can issue arbitrary
  HTTP methods (POST/PATCH/DELETE) so blanket-allowing it is unsafe.
- Leave kubectl/docker/aws/gcloud unclassified at the runner level; their
  surface area is too sprawling to classify safely without per-subcommand
  audits.
